backend/requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Firebase
firebase-admin==6.4.0

# AI - Gemini Vision
google-generativeai==0.3.2

# Image Processing
Pillow>=10.2.0

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0

# Telegram Bot (for Sprint 2+)
python-telegram-bot==20.7

# Development
pytest==7.4.4
httpx
pandas
